ongoing project
A project management to-do list is a collection of tasks that outlines the work a project manager or team plans to complete during a project. This list helps project managers understand the various tasks of a project so they can assign work to team members and fulfill their own duties.